Tap doesn't set hoverPseudoClass on an element in Android (close #294)
